Melhorado modelo de endereço
Melhorado modelo de endereço e criado enum para os estados brasileiros
