--- ifit/src/main/java/ifit/model/dto/Endereco.java
@@ -1,26 +1,0 @@
-package ifit.model.dto;
-
-public class Endereco {
-	String rua;
-	Integer numero;
-	String bairro;
-	
-	public String getRua() {
-		return rua;
-	}
-	public void setRua(String rua) {
-		this.rua = rua;
-	}
-	public Integer getNumero() {
-		return numero;
-	}
-	public void setNumero(Integer numero) {
-		this.numero = numero;
-	}
-	public String getBairro() {
-		return bairro;
-	}
-	public void setBairro(String bairro) {
-		this.bairro = bairro;
-	}
-}
