+++ ifit/src/main/java/ifit/model/endereco/Endereco.java
@@ -1,0 +1,43 @@
+package ifit.model.endereco;
+
+import ifit.model.endereco.enuns.EstadoEnum;
+
+public class Endereco {
+	
+	private String rua;
+	private Integer numero;
+	private String bairro;
+	private String cep;
+	private EstadoEnum estado;
+	
+	public String getRua() {
+		return rua;
+	}
+	public void setRua(String rua) {
+		this.rua = rua;
+	}
+	public Integer getNumero() {
+		return numero;
+	}
+	public void setNumero(Integer numero) {
+		this.numero = numero;
+	}
+	public String getBairro() {
+		return bairro;
+	}
+	public void setBairro(String bairro) {
+		this.bairro = bairro;
+	}
+	public String getCep() {
+		return cep;
+	}
+	public void setCep(String cep) {
+		this.cep = cep;
+	}
+	public EstadoEnum getEstado() {
+		return estado;
+	}
+	public void setEstado(EstadoEnum estado) {
+		this.estado = estado;
+	}
+}
