+++ ifit/src/main/java/ifit/model/endereco/enuns/EstadoEnum.java
@@ -1,0 +1,5 @@
+package ifit.model.endereco.enuns;
+
+public enum EstadoEnum {
+	AC, AL, AP, AM, BA, CE, DF, ES, GO, MA, MT, MS, MG, PA, PB, PR, PE, PI, RJ, RN, RS, RO, RR, SC, SP, SE, TO;
+}
